Fix several issues

+++ scripts/deploy.sh
@@ -1,17 +1,38 @@
 #!/bin/bash
+PACKAGES=$(grep "'" Cargo.toml | grep -v '#' | tr -d "'" | tr -d ",")
 
-PACKAGE_NAME=$(head Cargo.toml | awk '/^name/{print $3}' | tr -d '"' | tr -d "'")
-IMAGE_NAME=$(git remote get-url --all origin | cut -d ':' -f 2 | sed "s/\..*$//g")
+
 
 deploy() {
-	docker build -t ${IMAGE_NAME} .
+    for PACKAGE_NAME in ${PACKAGES}
+    do
+        if [ X"${PACKAGE_NAME}" != X"brickpack" ];then
+            IMAGE_NAME="afsec/${PACKAGE_NAME}"
+            echo
+            echo "Deploying ${IMAGE_NAME}..."
+            pushd ${PACKAGE_NAME}
+            docker build -t ${IMAGE_NAME} .
+            popd
+        fi
+    done
     exit 0
 }
 
 undeploy() {
-	docker rmi ${IMAGE_NAME}:latest
+    for PACKAGE_NAME in ${PACKAGES}
+    do
+        if [ X"${PACKAGE_NAME}" != X"brickpack" ];then
+            IMAGE_NAME="afsec/${PACKAGE_NAME}"
+            echo
+    	    echo "Undeploying ${IMAGE_NAME}..."
+            CONTAINERS=$((docker ps -a ; docker ps) | grep "${IMAGE_NAME}" | awk '{print $1}')
+            [ X"${CONTAINERS}" != X"" ] && docker rm ${CONTAINERS}
+	        docker rmi ${IMAGE_NAME}:latest
+        fi
+    done
     exit 0
 }
 
 
 [ X"${1}" == X"-u" ] && undeploy || deploy
+
